+++ src/PCLCrypto.Shared.Formatters/Formatters/KeyFormatter.cs
@@ -172,25 +172,6 @@ namespace PCLCrypto.Formatters
         internal RSAParameters Read(Stream stream)
         {
             var parameters = this.ReadCore(stream);
-
-#if DESKTOP
-            if (!CapiKeyFormatter.IsCapiCompatible(parameters))
-            {
-                // CAPI is faster than RSAManaged, so try to reformat the key
-                // such that CAPI can represent it, if possible.
-                // But some RSA keys (especially those generated by iOS devices)
-                // have their P component with more sig figs than their Q component,
-                // in which case CAPI just can't handle it.
-                // Only change the RSA parameters if we expect it will be useful,
-                // since changing them on iOS actually makes the app crash.
-                var trimmedParameters = NegotiateSizes(parameters);
-                if (CapiKeyFormatter.IsCapiCompatible(trimmedParameters))
-                {
-                    return trimmedParameters;
-                }
-            }
-#endif
-
             return parameters;
         }
