+++ src/PCLCrypto.Shared.Win32RSA/NCryptAsymmetricKeyProviderBase.cs
