+++ src/PCLCrypto.Shared.Win32RSA/NCryptKeyBase.cs
@@ -12,34 +12,36 @@ namespace PCLCrypto
     using System.Threading.Tasks;
     using PInvoke;
     using Validation;
-    using static PInvoke.BCrypt;
-    using Platform = Windows.Security.Cryptography.Core;
+    using static PInvoke.NCrypt;
 
     /// <summary>
-    /// The base class for BCrypt implementations of the <see cref="ICryptographicKey"/> interface.
+    /// The base class for NCrypt implementations of the <see cref="ICryptographicKey"/> interface.
     /// </summary>
-    internal abstract class BCryptKeyBase : CryptographicKey, ICryptographicKey
+    internal abstract class NCryptKeyBase : CryptographicKey, ICryptographicKey
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="BCryptKeyBase" /> class.
+        /// Initializes a new instance of the <see cref="NCryptKeyBase" /> class.
         /// </summary>
-        internal BCryptKeyBase()
+        /// <param name="key">The native NCrypt key that this instance represents.</param>
+        internal NCryptKeyBase(SafeKeyHandle key)
         {
+            Requires.NotNull(key, nameof(key));
+            this.Key = key;
         }
 
         /// <inheritdoc />
-        public int KeySize => BCryptGetProperty<int>(this.Key, PropertyNames.BCRYPT_KEY_LENGTH);
+        public int KeySize => NCryptGetProperty<int>(this.Key, KeyStoragePropertyIdentifiers.NCRYPT_LENGTH_PROPERTY);
 
         /// <summary>
-        /// Gets the handle to the BCrypt cryptographic key for purposes of key export.
+        /// Gets the handle to the NCrypt cryptographic key for purposes of key export.
         /// </summary>
-        protected abstract SafeKeyHandle Key { get; }
+        protected SafeKeyHandle Key { get; }
 
         /// <inheritdoc />
-        public abstract byte[] Export(CryptographicPrivateKeyBlobType blobType);
+        public abstract byte[] Export(CryptographicPrivateKeyBlobType blobType = CryptographicPrivateKeyBlobType.Pkcs8RawPrivateKeyInfo);
 
         /// <inheritdoc />
-        public abstract byte[] ExportPublicKey(CryptographicPublicKeyBlobType blobType);
+        public abstract byte[] ExportPublicKey(CryptographicPublicKeyBlobType blobType = CryptographicPublicKeyBlobType.X509SubjectPublicKeyInfo);
 
         /// <summary>
         /// Disposes of managed and native resources of this object.
