+++ src/PCLCrypto.WinRT/PCLCrypto.WinRT.csproj
@@ -41,18 +41,12 @@
     <Folder Include="Properties\" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="RsaKeyProvider.cs" />
     <Compile Include="ECDsaKeyProvider.cs" />
-    <Compile Include="CngUtilities.cs" />
     <Compile Include="CryptographicBuffer.cs" />
     <Compile Include="ECDiffieHellman.cs" />
     <Compile Include="ECDiffieHellmanCngPublicKeyFactory.cs" />
     <Compile Include="ECDiffieHellmanFactory.cs" />
     <Compile Include="ECDiffieHellmanPublicKey.cs" />
-    <Compile Include="BCryptKeyBase.cs" />
-    <Compile Include="NCryptAsymmetricKey.cs" />
-    <Compile Include="NCryptAsymmetricKeyProviderBase.cs" />
-    <Compile Include="NCryptKeyBase.cs" />
     <Compile Include="WinRTCryptographicKey.cs" />
     <Compile Include="WinRTCryptographicHash.cs" />
     <Compile Include="HashAlgorithmProvider.cs" />
@@ -89,4 +83,5 @@
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), EnlistmentInfo.targets))\EnlistmentInfo.targets" Condition=" '$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), EnlistmentInfo.targets))' != '' " />
   <Import Project="..\PCLCrypto.Shared.Common\PCLCrypto.Shared.Common.projitems" Label="Shared" Condition="Exists('..\PCLCrypto.Shared.Common\PCLCrypto.Shared.Common.projitems')" />
   <Import Project="..\PCLCrypto.Shared.PlatformCommon\PCLCrypto.Shared.PlatformCommon.projitems" Label="Shared" Condition="Exists('..\PCLCrypto.Shared.PlatformCommon\PCLCrypto.Shared.PlatformCommon.projitems')" />
+  <Import Project="..\PCLCrypto.Shared.Win32RSA\PCLCrypto.Shared.Win32RSA.projitems" Label="Shared" />
 </Project>
